requirements.txt
streamlit
pandas
numpy
plotly
sentence-transformers
google-generativeai
pyarrow
scikit-learn
networkx